refactor: 회원 관련 DTO 패키지 변경 및 inner record 변경
