+++ backend/ddang/src/main/java/com/ddang/ddang/user/presentation/UserController.java
@@ -5,9 +5,9 @@ import com.ddang.ddang.authentication.domain.dto.AuthenticationUserInfo;
 import com.ddang.ddang.image.presentation.util.ImageRelativeUrlFinder;
 import com.ddang.ddang.image.presentation.util.ImageTargetType;
 import com.ddang.ddang.user.application.UserService;
-import com.ddang.ddang.user.application.dto.ReadUserDto;
+import com.ddang.ddang.user.application.dto.request.ReadUserDto;
 import com.ddang.ddang.user.presentation.dto.response.ReadUserResponse;
-import com.ddang.ddang.user.application.dto.UpdateUserDto;
+import com.ddang.ddang.user.application.dto.response.UpdateUserDto;
 import com.ddang.ddang.user.presentation.dto.request.UpdateUserRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
