--- backend/ddang/src/main/java/com/ddang/ddang/user/presentation/dto/response/CategoryResponse.java
@@ -1,4 +1,0 @@
-package com.ddang.ddang.user.presentation.dto.response;
-
-public record CategoryResponse(String main, String sub) {
-}
