+++ backend/ddang/src/main/java/com/ddang/ddang/user/presentation/dto/response/ReadAuctionsResponse.java
@@ -1,5 +1,6 @@
 package com.ddang.ddang.user.presentation.dto.response;
 
+import com.ddang.ddang.auction.application.dto.response.ReadAuctionDto;
 import com.ddang.ddang.auction.application.dto.response.ReadAuctionsDto;
 
 import java.util.List;
@@ -17,5 +18,40 @@ public record ReadAuctionsResponse(List<ReadAuctionResponse> auctions, boolean i
                                .toList();
 
         return new ReadAuctionsResponse(readAuctionResponses, readAuctionsDto.isLast());
+    }
+
+    public record ReadAuctionResponse(
+            Long id,
+            String title,
+            String image,
+            int auctionPrice,
+            String status,
+            int auctioneerCount
+    ) {
+
+        public static ReadAuctionResponse of(final ReadAuctionDto dto, final String imageRelativeUrl) {
+            return new ReadAuctionResponse(
+                    dto.id(),
+                    dto.title(),
+                    calculateThumbnailImageUrl(dto, imageRelativeUrl),
+                    processAuctionPrice(dto.startPrice(), dto.lastBidPrice()),
+                    dto.auctionStatus().name(),
+                    dto.auctioneerCount()
+            );
+        }
+
+        private static String calculateThumbnailImageUrl(final ReadAuctionDto dto, final String imageRelativeUrl) {
+            final String thumbnailAuctionImageStoreName = dto.auctionImageStoreNames().get(0);
+
+            return imageRelativeUrl + thumbnailAuctionImageStoreName;
+        }
+
+        private static int processAuctionPrice(final Integer startPrice, final Integer lastBidPrice) {
+            if (lastBidPrice == null) {
+                return startPrice;
+            }
+
+            return lastBidPrice;
+        }
     }
 }
