+++ backend/ddang/src/main/java/com/ddang/ddang/user/presentation/dto/response/ReadUserResponse.java
@@ -1,6 +1,6 @@
 package com.ddang.ddang.user.presentation.dto.response;
 
-import com.ddang.ddang.user.application.dto.ReadUserDto;
+import com.ddang.ddang.user.application.dto.request.ReadUserDto;
 import com.ddang.ddang.user.presentation.util.ReliabilityProcessor;
 
 public record ReadUserResponse(String name, String profileImage, Float reliability) {
