--- backend/ddang/src/main/java/com/ddang/ddang/user/presentation/dto/response/SellerResponse.java
@@ -1,9 +1,0 @@
-package com.ddang.ddang.user.presentation.dto.response;
-
-public record SellerResponse(
-        Long id,
-        String image,
-        String nickname,
-        Float reliability
-) {
-}
